Cleanup ParamAttributeTypeConversions
This was formerly known as ParamNumberConversions, but that name didn't
accurately reflect what this class was for since it also handled some
non-numeric conversions. This change renames that class, along with the
associated test class, to be more descriptive, and fixes any
documentation comments that similarly alluded to this class only
supporting numeric conversions.

Also, the nunit copyright notice was added to this file and the
associated test file.
