+++ src/NUnitFramework/framework/Internal/ParamAttributeTypeConversions.cs
@@ -1,3 +1,26 @@
+// ***********************************************************************
+// Copyright (c) 2017 Charlie Poole
+//
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject to
+// the following conditions:
+// 
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+// ***********************************************************************
+
 using System;
 using System.Collections;
 using System.Reflection;
@@ -7,14 +30,14 @@ using NUnit.Framework.Interfaces;
 namespace NUnit.Framework.Internal
 {
     /// <summary>
-    /// Helper methods for converting parameters to numeric values
+    /// Helper methods for converting parameters to numeric values to supported types
     /// </summary>
-    internal static class ParamNumberConversions
+    internal static class ParamAttributeTypeConversions
     {
         /// <summary>
-        /// Converts an array of objects to the <paramref name="targetType"/>, if it is numeric.
+        /// Converts an array of objects to the <paramref name="targetType"/>, if it is supported.
         /// </summary>
-        public static IEnumerable ConvertDataToNumeric(object[] data, Type targetType)
+        public static IEnumerable ConvertData(object[] data, Type targetType)
         {
             if (targetType.GetTypeInfo().IsEnum && data.Length == 0)
             {
