+++ src/NUnitFramework/framework/nunit.framework-3.5.csproj
@@ -179,8 +179,8 @@
     <Compile Include="Internal\MethodWrapper.cs">
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="Internal\ParamAttributeTypeConversions.cs" />
     <Compile Include="Internal\ParameterWrapper.cs" />
-    <Compile Include="Internal\ParamNumberConversions.cs" />
     <Compile Include="Internal\TestCaseTimeoutException.cs" />
     <Compile Include="Internal\TestNameGenerator.cs" />
     <Compile Include="Internal\TypeNameDifferenceResolver.cs" />
