+++ src/NUnitFramework/tests/nunit.framework.tests-4.0.csproj
@@ -148,6 +148,7 @@
     <Compile Include="Internal\Filters\MockTestFilter.cs" />
     <Compile Include="Internal\NamespaceTreeBuilderTests.cs" />
     <Compile Include="Internal\Filters\MultipleFilterTests.cs" />
+    <Compile Include="Internal\ParamAttributeTypeConversionTests.cs" />
     <Compile Include="Internal\Results\AssertionResultTests.cs" />
     <Compile Include="Internal\Results\TestAttachmentTests.cs" />
     <Compile Include="Internal\Results\TestAttachmentXmlTests.cs" />
