+++ src/NUnitFramework/tests/nunit.framework.tests-netstandard16.csproj
@@ -220,6 +220,7 @@
     <Compile Include="Internal\Filters\TestFilterXmlTests.cs" />
     <Compile Include="Internal\Filters\TestNameFilterTests.cs" />
     <Compile Include="Internal\NamespaceTreeBuilderTests.cs" />
+    <Compile Include="Internal\ParamAttributeTypeConversionTests.cs" />
     <Compile Include="Internal\RealAsyncSetupTeardownTests.cs" />
     <Compile Include="Internal\AsyncSetupTeardownTests.cs" />
     <Compile Include="Internal\AsyncTestMethodTests.cs" />
